+++ cmd/list.go
@@ -20,11 +20,11 @@ import (
 
 	"github.com/kubicorn/kubicorn/pkg/cli"
 	"github.com/kubicorn/kubicorn/pkg/logger"
-	"github.com/kubicorn/kubicorn/state"
-	"github.com/kubicorn/kubicorn/state/fs"
-	"github.com/kubicorn/kubicorn/state/git"
-	"github.com/kubicorn/kubicorn/state/jsonfs"
-	"github.com/kubicorn/kubicorn/state/s3"
+	"github.com/kubicorn/kubicorn/pkg/state"
+	"github.com/kubicorn/kubicorn/pkg/state/fs"
+	"github.com/kubicorn/kubicorn/pkg/state/git"
+	"github.com/kubicorn/kubicorn/pkg/state/jsonfs"
+	"github.com/kubicorn/kubicorn/pkg/state/s3"
 	"github.com/minio/minio-go"
 	"github.com/spf13/cobra"
 )
