+++ pkg/cli/state_store.go
@@ -19,11 +19,11 @@ import (
 	"fmt"
 
 	"github.com/kubicorn/kubicorn/pkg/logger"
-	"github.com/kubicorn/kubicorn/state"
-	"github.com/kubicorn/kubicorn/state/fs"
-	"github.com/kubicorn/kubicorn/state/git"
-	"github.com/kubicorn/kubicorn/state/jsonfs"
-	"github.com/kubicorn/kubicorn/state/s3"
+	"github.com/kubicorn/kubicorn/pkg/state"
+	"github.com/kubicorn/kubicorn/pkg/state/fs"
+	"github.com/kubicorn/kubicorn/pkg/state/git"
+	"github.com/kubicorn/kubicorn/pkg/state/jsonfs"
+	"github.com/kubicorn/kubicorn/pkg/state/s3"
 	minio "github.com/minio/minio-go"
 	gg "github.com/tcnksm/go-gitconfig"
 )
