+++ pkg/state/fs/impl.go
@@ -25,7 +25,7 @@ import (
 	"github.com/ghodss/yaml"
 	"github.com/kubicorn/kubicorn/apis/cluster"
 	"github.com/kubicorn/kubicorn/pkg/logger"
-	"github.com/kubicorn/kubicorn/state"
+	"github.com/kubicorn/kubicorn/pkg/state"
 )
 
 type FileSystemStoreOptions struct {
