+++ pkg/state/git/impl_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package jsonfs
+package git
 
 import (
 	"encoding/json"
@@ -23,28 +23,33 @@ import (
 
 	"github.com/kubicorn/kubicorn/apis/cluster"
 	"github.com/kubicorn/kubicorn/profiles/amazon"
-	"github.com/kubicorn/kubicorn/state"
+	"github.com/kubicorn/kubicorn/pkg/state"
 )
 
-func TestJsonFileSystem(t *testing.T) {
+func TestJsonGit(t *testing.T) {
 	testFilePath := ".test/"
-	clusterName := "jsonfs-test"
+	clusterName := "git-test"
 	c := amazon.NewUbuntuCluster(clusterName)
-	o := &JSONFileSystemStoreOptions{
+	o := &JSONGitStoreOptions{
 		BasePath:    testFilePath,
 		ClusterName: c.Name,
+		CommitConfig: &JSONGitCommitConfig{
+			Name:   "Dummy Cluster",
+			Email:  "dummy@clustermail.co",
+			Remote: "https://github.com/kubicorn/kubicorn",
+		},
 	}
-	fs := NewJSONFileSystemStore(o)
-	if err := fs.Destroy(); err != nil {
+	git := NewJSONGitStore(o)
+	if err := git.Destroy(); err != nil {
 		t.Fatalf("Error destroying any existing state: %v", err)
 	}
-	if fs.Exists() {
+	if git.Exists() {
 		t.Fatalf("State shouldn't exist because we just destroyed it, but Exists() returned true")
 	}
-	if err := fs.Commit(c); err != nil {
+	if err := git.Commit(c); err != nil {
 		t.Fatalf("Error committing cluster: %v", err)
 	}
-	files, err := fs.List()
+	files, err := git.List()
 	if err != nil {
 		t.Fatalf("Error listing files: %v", err)
 	}
@@ -54,7 +59,7 @@ func TestJsonFileSystem(t *testing.T) {
 	if files[0] != state.ClusterJSONFile {
 		t.Fatalf("Expected file name to be %v, got %v", state.ClusterJSONFile, files[0])
 	}
-	read, err := fs.GetCluster()
+	read, err := git.GetCluster()
 	if err != nil {
 		t.Fatalf("Error getting cluster: %v", err)
 	}
@@ -72,7 +77,7 @@ func TestJsonFileSystem(t *testing.T) {
 	if !reflect.DeepEqual(unmarshalled, c) {
 		t.Fatalf("Cluster read directly from json file doesn't equal cluster inputted: %v", unmarshalled)
 	}
-	if err = fs.Destroy(); err != nil {
+	if err = git.Destroy(); err != nil {
 		t.Fatalf("Error cleaning up state: %v", err)
 	}
 }
