+++ pkg/state/interface.go
