+++ pkg/state/s3/impl.go
@@ -23,7 +23,7 @@ import (
 
 	"github.com/kubicorn/kubicorn/apis/cluster"
 	"github.com/kubicorn/kubicorn/pkg/logger"
-	"github.com/kubicorn/kubicorn/state"
+	"github.com/kubicorn/kubicorn/pkg/state"
 	"github.com/minio/minio-go"
 )
